% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_graph.R
\name{nn_graph}
\alias{nn_graph}
\title{Create a diagram with \code{nomnoml} based on edges and nodes}
\usage{
nn_graph(
  nodes,
  edges,
  output = "diagram",
  direction = "down",
  lineWidth = 1,
  fill = "#FEFEFF",
  zoom = 4,
  arrowSize = 1,
  bendSize = 0.3,
  gutter = 5,
  edgeMargin = 0,
  edgesStyle = "hard",
  fillArrows = "false",
  font = "sans",
  fontSize = 12,
  leading = 1.25,
  padding = 8,
  spacing = 40,
  stroke = "#33322E",
  title = "filename",
  png = NULL,
  width = NULL,
  height = NULL,
  svg = FALSE
)
}
\arguments{
\item{nodes}{A data frame with two columns: \code{id}, \code{text}, and an optional
\code{classifier}.}

\item{edges}{A data frame with three columns: \code{from}, \code{to}, and
\code{association}.}

\item{output}{Defaults to "diagram". Valid values are either "diagram" or
"code".}

\item{direction}{Defaults to "down". Valid values are either "down" or
"right".}

\item{fill}{Defaults to "#FEFEFF". Values commonly used by \code{nomnoml} inlcude
"#eee8d5" and "#fdf6e3".}

\item{edgesStyle}{Defaults to "hard". Valid values are either "hard" or
"rounded". Value is passed to \code{nomnoml} as "edges".}

\item{png}{Optional file name to export diagram as 'png'.}

\item{width}{Optional width in pixels for the exported 'png'.}

\item{height}{Optional height in pixels for the exported 'png'.}

\item{svg}{Use 'svg' output instead of 'png'? Notice that rendering in
'svg' is not at a par with 'png' and renders incorrectly at times.}
}
\value{
An object with class \code{nomnoml} and \code{htmlwidget}. By default, it the
diagram is printed.
}
\description{
Available \code{association} types to be used in the \code{edges} data frame:
}
\details{
\itemize{
\item association
->   association
<->  association
-->  dependency
<--> dependency
-:>  generalization
<:-  generalization
--:> implementation
<:-- implementation
+-   composition
+->  composition
o-   aggregation
o->  aggregation
--   note
-/-  hidden
}
}
\examples{
nodes <- tibble::tribble(
  ~id, ~text,
  1, "Starting point",
  2, "Other starting point",
  3, "This is where things merge",
  4, "This is where things go"
)


edges <- tibble::tribble(
  ~from, ~to, ~association,
  1, 3, "-",
  2, 3, "-",
  3, 4, "->"
)

nn_graph(
  nodes = nodes,
  edges = edges,
  fill = "#fdf6e3",
  edgesStyle = "rounded",
  # font = "Roboto Condensed",
  fillArrows = "false",
  svg = TRUE
)
}
